Silent SSO
Don't ask for confirmation when user already logged in Auth0 SSO.

By passing `?prompt=none` in the `/authorize` request Auth0 will not
ask to confirm the identity when already logged in SSO.

The effect will be that the user will be transparently be logged in the
application (RStudio in this case).

NOTE: When `?prompt=none` is passed to `/authorize` but the user is *not*
logged in SSO the user will not be promped to login and therefore can't
be logged in. In this case the application redirects to `/login` again by
passing `?prompt=true` so that the user can login.

SEE: https://auth0.com/docs/api-auth/tutorials/silent-authentication