Run websocket connections through express middleware
Similar to the fix for the generic auth-proxy:
https://github.com/ministryofjustice/analytics-platform-auth-proxy but in
addition it applied an authentication check to the websocket. This is
non-standard as a normal application would use the header sec-websocket-key to
ensure the incoming websocket connection was solicited, but rstudio relies on a
cookie header so we need to ensure that someone is authenticated and authorised
before injecting the cookie header to the downstream rstudio server.