Don't read body, just set the `Content-Length` header
We used to read the request body to calculate the `Content-Length`,
we now just copy the velue of the `Content-Length` header.
The request body is proxied by default, as you'd expect.

Tested locally against a running RStudio container and it seems to
work. I've uploaded a 37MB ISO image and checked the MD5 of the file
uploaded against the checksum of the original file and the file was
uploaded without corruption.

**NOTE**: This is to simplify the `rstudio-proxy` logic before moving
into the other `auth-proxy`.

Part of ticket: https://trello.com/c/5twZHrb4/215-try-to-replace-custom-rstudio-auth-proxy-with-our-usual-generic-one